+++ nonebot/typing.py
@@ -1,9 +1,11 @@
-from typing import TYPE_CHECKING, Union, List, Dict, Any, Sequence, Callable, Tuple, Awaitable, Pattern, Iterable
+from typing import TYPE_CHECKING, Optional, Union, List, Dict, Any, Sequence, Callable, Tuple, Awaitable, Pattern, Iterable
 
 if TYPE_CHECKING:
     from aiocqhttp import Event as CQEvent
     from nonebot import NoneBot
     from nonebot.command import CommandSession
+    from nonebot.natural_language import NLPSession, IntentCommand
+    from nonebot.notice_request import NoticeSession, RequestSession
 
 Context_T = Dict[str, Any]
 Message_T = Union[str, Dict[str, Any], List[Dict[str, Any]]]
@@ -15,6 +17,9 @@ Patterns_T = Union[Iterable[str], str, Iterable[Pattern[str]], Pattern[str]]
 State_T = Dict[str, Any]
 Filter_T = Callable[[Any], Union[Any, Awaitable[Any]]]
 PermChecker_T = Callable[["NoneBot", "CQEvent"], Awaitable[bool]]
+NLPHandler_T = Callable[["NLPSession"], Awaitable[Optional["IntentCommand"]]]
+NoticeHandler_T = Callable[["NoticeSession"], Awaitable[Any]]
+RequestHandler_T = Callable[["RequestSession"] , Awaitable[Any]]
 
 
 __all__ = [